Fixed a bootup error

+++ run.py
@@ -273,6 +273,8 @@ class BattleBananaClient(discord.AutoShardedClient):
                 player.save()
 
     async def on_guild_remove(self, guild):
+        if not self.is_ready():
+            return
         for collection in dbconn.db.list_collection_names():
             if collection not in ("Player", "Topdogs"):
                 dbconn.db[collection].delete_many({'_id': {'$regex': '%s.*' % guild.id}})
